Add simple roast bot
[ghstack-poisoned]
