+++ .envrc
@@ -6,3 +6,5 @@ if has nix; then
   # Use nix flakes to setup the environment
   use flake . --no-pure-eval --accept-flake-config
 fi
+
+dotenv
