+++ apeiron/web/app.py
@@ -1,4 +1,5 @@
 import logging
+import os
 
 from fastapi import FastAPI
 
@@ -7,8 +8,7 @@ from . import discord, github
 logging.basicConfig(level=logging.DEBUG)
 
 
-app = FastAPI(title="Apeiron Discord Webhook API")
-logger = logging.getLogger(__name__)
+app = FastAPI(title="Apeiron Webhook API")
 
 
 @app.get("/health")
@@ -17,11 +17,11 @@ async def health_check():
 
 
 app.include_router(
-    discord.router,
+    discord.create_router(os.getenv("DISCORD_PUBLIC_KEY")),
     tags=["discord"],
 )
 
 app.include_router(
-    github.router,
+    github.create_router(os.getenv("GITHUB_WEBHOOK_SECRET")),
     tags=["github"],
 )
