+++ pyproject.toml
@@ -4,7 +4,16 @@ version = "0.1.0"
 description = "A multitask AI agent designed to streamline and automate operational workflows"
 readme = "README.md"
 requires-python = ">=3.13"
-dependencies = ["fastapi[standard]>=0.110.0", "pynacl>=1.5.0"]
+dependencies = [
+  "click>=8.1.8",
+  "fastapi[standard]>=0.110.0",
+  "pynacl>=1.5.0",
+  "openai>=1.12.0",
+  "langchain-mistralai>=0.2.7",
+  "langchain-core>=0.3.40",
+  "discord>=2.3.2",
+  "langgraph>=0.3.2",
+]
 
 [tool.ruff.lint]
 select = [
